rename module and use asset spec instead
