+++ python_modules/libraries/dagster-embedded-elt/README.md
@@ -1,4 +1,4 @@
 # dagster-embedded-elt
 
-The docs for `dagster-elt` can be found
+The docs for `dagster-embedded-elt` can be found
 [here](https://docs.dagster.io/_apidocs/libraries/dagster-embedded-elt).
