--- python_modules/libraries/dagster-embedded-elt/dagster_embbeded_elt/sling/__init__.py
@@ -1,11 +1,0 @@
-from dagster_elt.sling.asset_defs import build_sling_asset
-from dagster_elt.sling.resources import (
-    SlingResource,
-    SlingMode,
-)
-
-__all__ = [
-    "SlingResource",
-    "SlingMode",
-    "build_sling_asset",
-]
