+++ python_modules/libraries/dagster-embedded-elt/setup.cfg
@@ -6,4 +6,4 @@ ignore =
     .coveragerc
     tox.ini
     pytest.ini
-    dagster_elt_tests/**
+    dagster_embedded_elt_tests/**
