+++ python_modules/libraries/dagster-embedded-elt/setup.py
@@ -6,7 +6,7 @@ from setuptools import find_packages, setup
 
 def get_version() -> str:
     version: Dict[str, str] = {}
-    with open(Path(__file__).parent / "dagster_elt/version.py", encoding="utf8") as fp:
+    with open(Path(__file__).parent / "dagster_embedded_elt/version.py", encoding="utf8") as fp:
         exec(fp.read(), version)
 
     return version["__version__"]
@@ -16,7 +16,7 @@ ver = get_version()
 # dont pin dev installs to avoid pip dep resolver issues
 pin = "" if ver == "1!0+dev" else f"=={ver}"
 setup(
-    name="dagster-elt",
+    name="dagster-embedded-elt",
     version=ver,
     author="Dagster Labs",
     author_email="hello@dagsterlabs.com",
